refactor: split recherche offres tests
